feat(#14,#15,#16,#17,#18): floats, operadorAtendimento e estatisticas por operador
- tempoCompra e tempoPassagem convertidos de int para float em toda a
  codebase (produto, cliente, caixa, snapshot, catalog) — remove minimo
  artificial de 2 e preserva valores do ficheiro sem truncagem
- #15: adicionar operadorAtendimento ao Cliente, preenchido em
  concluir_atendimento com o operador ativo; historico_caixas.csv usa
  operadorAtendimento por cliente em vez de ultimoOperador da caixa
- #16: calcular_estatisticas_operadores agrega historicoAtendidos por
  operadorAtendimento on-demand; mostrar_estatisticas_finais mostra lista
  por operador e minimo real; guardar_estatisticas_csv tem secao por operador
- #17: abrir_log_acoes faz append se ficheiro existe, cria com cabecalho
  se nao existe; main regista INICIO_SESSAO e FIM_SESSAO
- #18: totalValorVendido no Supermercado acumulado em concluir_atendimento;
  exportado em estatisticas.csv e mostrado nas estatisticas finais

+++ src/product.h
@@ -7,13 +7,13 @@ typedef struct {
     int id;
     char nome[MAX_NOME];
     float preco;
-    int tempoCompra;
-    int tempoPassagem;
+    float tempoCompra;
+    float tempoPassagem;
     bool oferecido;
 } Produto;
 
-int tempo_total_produtos(const Produto *produtos, int nProdutos);
-int tempo_compra_total_produtos(const Produto *produtos, int nProdutos);
+float tempo_total_produtos(const Produto *produtos, int nProdutos);
+float tempo_compra_total_produtos(const Produto *produtos, int nProdutos);
 float valor_total_produtos(const Produto *produtos, int nProdutos);
 void libertar_produtos(Produto *produtos);
